+++ tests/testthat/test-other.R
@@ -76,7 +76,7 @@ test_that("ec.snip", {
   expect_equal(length(op$series), 3)
 })
 
-test_that("ec.data for dendrogram", {
+test_that("ec.data format dendrogram", {
   hc <- hclust(dist(USArrests), "average")
   p <- ec.init(preset= FALSE)
   p$x$opts$series <- list(list(
@@ -89,11 +89,82 @@ test_that("ec.data for dendrogram", {
   expect_equal(length(p$x$opts$series[[1]]$data[[1]]$children[[1]]$children), 2)
 })
 
-test_that("ec.data for boxlpot", {
+test_that("ec.data format boxlpot", {
   ds <- mtcars |> dplyr::relocate(am,mpg) |> ec.data(format='boxplot')
   expect_equal(ds$series[[1]]$type, 'boxplot')
   expect_equal(ds$dataset$source[[1]], c("V1","V2","V3","V4","V5","V6"))
   expect_equal(class(ds$axlbl), 'JS_EVAL')
+})
+
+test_that("ec.data for treePC", {
+  df <- as.data.frame(Titanic) |> group_by(Survived,Class) |> 
+    summarise(value=sum(Freq), .groups='drop') |>
+    mutate(parents= as.character(Survived), 
+           children= as.character(Class)) |>
+    select(parents, children, value)
+  # add root to form a tree
+  df[nrow(df) + 1,] <- list('survived','Yes',711)
+  df[nrow(df) + 1,] <- list('survived','No', 1490)
+  df[nrow(df) + 1,] <- list('root2','survived',2201)
+  
+  p <- ec.init(preset=FALSE)
+  p$x$opts$series <- list(list(
+    type='sunburst', 
+    data= ec.data(df, format='treePC')[[1]]$children, 
+    radius=c('11%', '90%')
+    #,label=list(rotate='radial'), emphasis=list(focus='none')
+  ))
+  
+  expect_equal(p$x$opts$series[[1]]$data[[1]]$value, 711)
+  expect_equal(length(p$x$opts$series[[1]]$data[[1]]$children), 4)
+})
+
+test_that("ec.data format treeTK", {
+  df <- as.data.frame(Titanic) |> 
+    group_by(Survived,Age,Sex,Class) |>
+    summarise(value=sum(Freq), .groups='drop') |>
+    mutate(pathString= paste('Survive?', Survived, Age, Sex, Class, sep='/'),
+           itemStyle= case_when(Survived=='Yes' ~ "color='green'", 
+                                TRUE ~ "color='pink'"))  |> 
+    select(pathString,value,itemStyle) |>
+    rbind( as.data.frame(Titanic) |> 
+             group_by(Survived,Age,Sex) |>
+             summarise(value=sum(Freq), .groups='drop') |>
+             mutate(pathString= paste('Survive?', Survived, Age, Sex, sep='/'),
+                    itemStyle= case_when(Survived=='Yes' ~ "color='green'", 
+                                         TRUE ~ "color='pink'"))  |> 
+             select(pathString,value,itemStyle)) |>
+    rbind( as.data.frame(Titanic) |> 
+             group_by(Survived,Age) |> 
+             summarise(value=sum(Freq), .groups='drop') |>
+             mutate(pathString= paste('Survive?', Survived, Age, sep='/'),
+                    itemStyle= case_when(Survived=='Yes' ~ "color='green'", 
+                                         TRUE ~ "color='pink'"))  |> 
+             select(pathString,value,itemStyle)) |>
+    rbind( as.data.frame(Titanic) |> 
+             group_by(Survived) |>
+             summarise(value=sum(Freq), .groups='drop') |>
+             mutate(pathString= paste('Survive?', Survived, sep='/'),
+                    itemStyle= case_when(Survived=='Yes' ~ "color='green'", 
+                                         TRUE ~ "color='pink'"))  |> 
+             select(pathString,value,itemStyle))
+  dat1 <- ec.data(df, format='treeTK')
+  dat1[[1]]$itemStyle <- list(color= 'white')
+  
+  p <- ec.init(preset=FALSE, title=list(text= 'Titanic: Survival by Class'))
+  p$x$opts$series <- list(list(
+    type= 'sunburst', radius= c(0, '90%'), label=list(rotate=0),
+    #type= 'tree', symbolSize= ec.clmn(scale=0.08),
+    #type= 'treemap', upperLabel= list(show=TRUE, height=30), itemStyle= list(borderColor= '#999'), #leafDepth=4,
+    data= dat1, 
+    #	levels= list(list(itemStyle= list(color= 'blue')),
+    #					 list(itemStyle= list(color= 'pink'))),
+    emphasis= list(focus='none') 
+  ))
+  p$x$opts$tooltip <- list(show=TRUE, formatter=ec.clmn('%@%','pct')) # OK
+  
+  expect_equal(p$x$opts$series[[1]]$data[[1]]$value, 2201)
+  expect_equal(length(p$x$opts$series[[1]]$data[[1]]$children), 2)
 })
 
 test_that("load 3D surface", {
@@ -108,3 +179,21 @@ test_that("load 3D surface", {
   }
   else expect_equal(1,1)
 })
+
+test_that("shapefiles with multi-polygons", {
+  if (interactive()) {
+    library(sf)
+    fname <- system.file("shape/nc.shp", package="sf")
+    nc <- st_read(fname)
+    p <- ec.init(load= c('leaflet', 'custom'),  # load custom for polygons
+           js= ec.util(type= 'sf.bbox', bbox= st_bbox(nc)),
+           series= ec.util(df= nc, nid= 'NAME', itemStyle= list(opacity= 0.3)),
+           tooltip= list(show= TRUE, formatter= '{a}')
+    )
+    expect_true(p$x$opts$leaflet$roam)
+    expect_equal(p$x$opts$series[[108]]$name, 'Brunswick')
+    expect_equal(p$x$opts$series[[108]]$itemStyle$opacity, 0.3)
+  }
+  else expect_equal(1,1)
+})
+